COMMERCE-4987 commerce-machine-learning-impl Move out entry class pk from common classes
