+++ modules/dxp/apps/commerce/commerce-machine-learning-impl/src/main/java/com/liferay/commerce/machine/learning/internal/recommendation/model/UserCommerceMLRecommendationImpl.java
@@ -29,10 +29,21 @@ public class UserCommerceMLRecommendationImpl
 	}
 
 	@Override
+	public long getEntryClassPK() {
+		return _entryClassPK;
+	}
+
+	@Override
 	public void setAssetCategoryIds(long[] assetCategoryIds) {
 		_assetCategoryIds = assetCategoryIds;
 	}
 
+	@Override
+	public void setEntryClassPK(long entryClassPK) {
+		_entryClassPK = entryClassPK;
+	}
+
 	private long[] _assetCategoryIds;
+	private long _entryClassPK;
 
 }
